move function in transition component to helper, and add warning when user use the wrong transition component
