+++ src/victory-transition/victory-transition.js
@@ -1,7 +1,7 @@
 import React from "react";
 import VictoryAnimation from "../victory-animation/victory-animation";
-import { Transitions } from "../victory-util/index";
-import { defaults, isFunction, pick } from "lodash";
+import { Transitions, TransitionHelpers } from "../victory-util/index";
+import { defaults, pick } from "lodash";
 
 export default class VictoryTransition extends React.Component {
   static propTypes = {
@@ -51,29 +51,6 @@ export default class VictoryTransition extends React.Component {
     }
   }
 
-  getDomainFromChildren(props, axis) {
-    const getChildDomains = (children) => {
-      return children.reduce((memo, child) => {
-        if (child.type && isFunction(child.type.getDomain)) {
-          const childDomain = child.props && child.type.getDomain(child.props, axis);
-          return childDomain ? memo.concat(childDomain) : memo;
-        } else if (child.props && child.props.children) {
-          return memo.concat(getChildDomains(React.Children.toArray(child.props.children)));
-        }
-        return memo;
-      }, []);
-    };
-
-    const childComponents = React.Children.toArray(props.children);
-    if (props.domain && (Array.isArray(props.domain) || props.domain[axis])) {
-      return Array.isArray(props.domain) ? props.domain : props.domain[axis];
-    } else {
-      const childDomains = getChildDomains(childComponents);
-      return childDomains.length === 0 ?
-        [0, 1] : [Math.min(...childDomains), Math.max(...childDomains)];
-    }
-  }
-
   render() {
     const props = this.state && this.state.nodesWillExit ?
       this.state.oldProps : this.props;
@@ -84,11 +61,21 @@ export default class VictoryTransition extends React.Component {
         this.state,
         (newState) => this.setState(newState)
       );
+
+    // whitelist using continuous-transition, otherwise give warning.
+    // eslint-disable-next-line max-len
+    const whitelist = ["VictoryBar", "VictoryAxis", "VictoryCandlestick", "VictoryErrorBar", "VictoryScatter"];
     const child = React.Children.toArray(props.children)[0];
+
+    if (whitelist.indexOf(child.type.displayName) === -1) {
+      // eslint-disable-next-line no-console, max-len, no-undef
+      console.warn("Component VictoryTransition should work with discrete charts, for continuous charts use VictoryContinuousTransition instead!");
+    }
+
     const transitionProps = getTransitionProps(child);
     const domain = {
-      x: this.getDomainFromChildren(props, "x"),
-      y: this.getDomainFromChildren(props, "y")
+      x: TransitionHelpers.getDomainFromChildren(props, "x"),
+      y: TransitionHelpers.getDomainFromChildren(props, "y")
     };
     const combinedProps = defaults(
       {domain}, transitionProps, child.props
