+++ src/victory-util/index.js
@@ -6,6 +6,7 @@ import PropTypes from "./prop-types";
 import Events from "./events";
 import * as Transitions from "./transitions";
 import * as ContinuousTransitions from "./continuous-transitions";
+import * as TransitionHelpers from "./transition-helpers";
 
 export {
   Collection,
@@ -14,6 +15,7 @@ export {
   Style,
   Transitions,
   ContinuousTransitions,
+  TransitionHelpers,
   PropTypes,
   Events
 };
